add migrations

+++ database/migrations/2022_02_06_071635_create_school_students_table.php
@@ -15,9 +15,9 @@ class CreateSchoolStudentsTable extends Migration
     {
         Schema::create('school_students', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('student_id')->nullable();
+            $table->integer('student_id')->nullable();
             $table->foreign('student_id')->references('id')->on('students')->onDelete('cascade');
-            $table->unsignedBigInteger('school_id')->nullable();
+            $table->integer('school_id')->nullable();
             $table->foreign('school_id')->references('id')->on('schools')->onDelete('cascade');
             $table->timestamps();
             $table->softDeletes();
